requirements.txt
Pillow>=10.0.0
FastAPI>=0.116.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.20
PySide6>=6.8.0
httpx>=0.27.0
